pyproject.toml
[tool.poetry]
name = "ml-tech-assessment"
version = "1.0.0"
description = "Transcript Analysis API - Analyzes transcripts and returns summaries with action items"
authors = ["jhonvalderrama <jhonvalderrama@aceup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.76.2"
pydantic-settings = "^2.9.1"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
